automation/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mindonroad-automation"
version = "0.2.0"
description = "Automation and document intelligence scripts for MindOnRoad."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "paddlepaddle==3.2.0",
  "paddleocr==3.3.3",
  "Pillow>=10.3.0",
  "pydantic>=2.8.0",
  "PyMuPDF>=1.24.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.29.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
extract-bg-document = "document_intelligence.cli:main"
ocr-worker = "document_intelligence.worker:main"
